requirements.txt
pandas
psycopg2
sqlalchemy
sqlalchemy
requests
paramiko
flask_testing
flask_restful
dataset
pytest


